Remove merlin-server as a dependency of orchestration-utils
