+++ merlin-driver/src/main/java/gov/nasa/jpl/aerie/merlin/driver/timeline/EventGraphFlattener.java
@@ -1,6 +1,5 @@
-package gov.nasa.jpl.aerie.merlin.server.remotes.postgres;
+package gov.nasa.jpl.aerie.merlin.driver.timeline;
 
-import gov.nasa.jpl.aerie.merlin.driver.timeline.EventGraph;
 import gov.nasa.jpl.aerie.merlin.protocol.model.EffectTrait;
 import org.apache.commons.lang3.tuple.Pair;
