+++ merlin-server/src/main/java/gov/nasa/jpl/aerie/merlin/server/remotes/postgres/PostgresParsers.java
@@ -28,49 +28,11 @@ import java.util.Map;
 import static gov.nasa.jpl.aerie.json.BasicParsers.*;
 import static gov.nasa.jpl.aerie.json.Uncurry.tuple;
 import static gov.nasa.jpl.aerie.json.Uncurry.untuple;
+import static gov.nasa.jpl.aerie.merlin.driver.json.Parsers.activityArgumentsP;
 import static gov.nasa.jpl.aerie.merlin.driver.json.SerializedValueJsonParser.serializedValueP;
 import static gov.nasa.jpl.aerie.merlin.driver.json.ValueSchemaJsonParser.valueSchemaP;
 
 public final class PostgresParsers {
-
-  public static final JsonParser<Timestamp> pgTimestampP = new JsonParser<>() {
-    private static final DateTimeFormatter format =
-        new DateTimeFormatterBuilder()
-            .append(DateTimeFormatter.ISO_OFFSET_DATE_TIME)
-            .appendFraction(ChronoField.MICRO_OF_SECOND, 0, 6, true)
-            .toFormatter();
-
-    @Override
-    public JsonObject getSchema(final SchemaCache anchors) {
-      return Json
-          .createObjectBuilder(stringP.getSchema())
-          .add("format", "date-time")
-          .build();
-    }
-
-    @Override
-    public JsonParseResult<Timestamp> parse(final JsonValue json) {
-      final var result = stringP.parse(json);
-      if (result instanceof final JsonParseResult.Success<String> s) {
-        try {
-          final var instant = LocalDateTime.parse(s.result(), format).atZone(ZoneOffset.UTC);
-          return JsonParseResult.success(new Timestamp(instant));
-        } catch (final DateTimeParseException e) {
-          return JsonParseResult.failure("invalid timestamp format "+e);
-        }
-      } else if (result instanceof final JsonParseResult.Failure<?> f) {
-        return f.cast();
-      } else {
-        throw new UnexpectedSubtypeError(JsonParseResult.class, result);
-      }
-    }
-
-    @Override
-    public JsonValue unparse(final Timestamp value) {
-      final var s = format.format(value.toInstant().atZone(ZoneOffset.UTC));
-      return stringP.unparse(s);
-    }
-  };
 
   public static final JsonParser<Pair<String, ValueSchema>> discreteProfileTypeP =
       productP
@@ -117,9 +79,6 @@ public final class PostgresParsers {
   }
 
   public static final JsonParser<Map<String, SerializedValue>> constraintArgumentsP = mapP(serializedValueP);
-  public static final JsonParser<Map<String, SerializedValue>> activityArgumentsP = mapP(serializedValueP);
-
-  public static final JsonParser<Map<String, SerializedValue>> simulationArgumentsP = mapP(serializedValueP);
 
   public static final JsonParser<ActivityAttributesRecord> activityAttributesP = productP
       .optionalField("directiveId", longP)
